unneeded float command

+++ worlds/frogmonster/__init__.py
@@ -161,7 +161,7 @@ class FrogmonsterWorld(World):
         slot_data["shuffled_bug_effects"] = bug_effect_array
 
         # Other Options:
-        slot_data["shop_multiplier"] = float(self.options.shop_multiplier.value / 100) # Convert to decimal for client
+        slot_data["shop_multiplier"] = self.options.shop_multiplier.value / 100 # Convert to decimal for client
         slot_data["shuffle_puzzles"] = bool(self.options.shuffle_puzzles.value)
         slot_data["open_city"] = bool(self.options.open_city.value)
         slot_data["death_link"] = bool(self.options.death_link.value)
